docs: 📜 update readme
this commit updates the readme to reflect the steps needed to run the
core, staging, and storage provider services locally alongside the
frontend web interface.

this fixes some typos, adds previously missing steps, and rephrases the
shell snippets in terms of the Makefile commands that were introduced in
pr #246.

a note that this can be done automatically using the `reset_env` script
has also been moved up to the top of the document, so that new
contributors don't step through the steps manually before discovering
the automatic way to perform this setup.

Refs: #246