ci: 📴 disable failing `build_containers` job
the `build_containers` CI job, run when a commit is pushed to the `main`
branch, has been broken for a few months. this prevents other jobs like
commit linting and tests from running after a PR is merged.

this deletes the `build_containers.sh` script and the corresponding
`containers.yml` workflow file. we can reintroduce this when we are
ready to be building containers once more.