docs: 💘 document how to work with databases
as part of my work on ENG-468, i found myself investigating how to (a)
connect to the databases for each of the Banyan services, (b) see the
tables and indexes within each database, and (c) see the schema of a
particular table.

this adds documentation to the README explaining how to connect to the
databases, and adds three simple makefile commands to help make this
knowledge accessible and discoverable to others.