chore: 🎴 portable `prepare_queries` scripts
invoke `env` to find the user's bash shell, in case it does not live in
the expected place. i.e. on my machine, bash is found at...

```
; which bash
/run/current-system/sw/bin/bash
```

for similar reasons, we `cargo sqlx`, in case `.cargo/bin/` is not part
of somebody's global `$PATH`.

finally, some trailing empty lines are removed. that part is cosmetic.